+++ lua/neotest-golang/discover_positions.lua
@@ -1,0 +1,100 @@
+local lib = require("neotest.lib")
+
+local M = {}
+
+function M.discover_positions(file_path)
+  local functions_and_methods = [[
+    ;;query
+    ((function_declaration
+      name: (identifier) @test.name)
+      (#match? @test.name "^(Test|Example)"))
+      @test.definition
+
+    (method_declaration
+      name: (field_identifier) @test.name
+      (#match? @test.name "^(Test|Example)")) @test.definition
+
+    (call_expression
+      function: (selector_expression
+        field: (field_identifier) @test.method)
+        (#match? @test.method "^Run$")
+      arguments: (argument_list . (interpreted_string_literal) @test.name))
+      @test.definition
+  ]]
+
+  local table_tests = [[
+    ;; query for list table tests
+        (block
+          (short_var_declaration
+            left: (expression_list
+              (identifier) @test.cases)
+            right: (expression_list
+              (composite_literal
+                (literal_value
+                  (literal_element
+                    (literal_value
+                      (keyed_element
+                        (literal_element
+                          (identifier) @test.field.name)
+                        (literal_element
+                          (interpreted_string_literal) @test.name)))) @test.definition))))
+          (for_statement
+            (range_clause
+              left: (expression_list
+                (identifier) @test.case)
+              right: (identifier) @test.cases1
+                (#eq? @test.cases @test.cases1))
+            body: (block
+             (expression_statement
+              (call_expression
+                function: (selector_expression
+                  field: (field_identifier) @test.method)
+                  (#match? @test.method "^Run$")
+                arguments: (argument_list
+                  (selector_expression
+                    operand: (identifier) @test.case1
+                    (#eq? @test.case @test.case1)
+                    field: (field_identifier) @test.field.name1
+                    (#eq? @test.field.name @test.field.name1))))))))
+
+    ;; query for map table tests 
+      (block
+          (short_var_declaration
+            left: (expression_list
+              (identifier) @test.cases)
+            right: (expression_list
+              (composite_literal
+                (literal_value
+                  (keyed_element
+                  (literal_element
+                      (interpreted_string_literal)  @test.name)
+                    (literal_element
+                      (literal_value)  @test.definition))))))
+        (for_statement
+           (range_clause
+              left: (expression_list
+                ((identifier) @test.key.name)
+                ((identifier) @test.case))
+              right: (identifier) @test.cases1
+                (#eq? @test.cases @test.cases1))
+            body: (block
+               (expression_statement
+                (call_expression
+                  function: (selector_expression
+                    field: (field_identifier) @test.method)
+                    (#match? @test.method "^Run$")
+                    arguments: (argument_list
+                    ((identifier) @test.key.name1
+                    (#eq? @test.key.name @test.key.name1))))))))
+  ]]
+
+  local query = functions_and_methods .. table_tests
+  local opts = { nested_tests = true }
+
+  ---@type neotest.Tree
+  local positions = lib.treesitter.parse_positions(file_path, query, opts)
+
+  return positions
+end
+
+return M
